// frontend/src/api/generated/models/alertRuleResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */
import type { AlertRuleResponseRuleType } from './alertRuleResponseRuleType';
import type { AlertRuleResponseSeverity } from './alertRuleResponseSeverity';
import type { AlertRuleResponseConditions } from './alertRuleResponseConditions';

export interface AlertRuleResponse {
  id: number;
  name: string;
  description: string;
  ruleType: AlertRuleResponseRuleType;
  severity: AlertRuleResponseSeverity;
  conditions: AlertRuleResponseConditions;
  enabled: boolean;
  createdAt: string;
  updatedAt: string;
}