// frontend/src/api/generated/models/createAlertRuleDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */
import type { CreateAlertRuleDtoRuleType } from './createAlertRuleDtoRuleType';
import type { CreateAlertRuleDtoSeverity } from './createAlertRuleDtoSeverity';
import type { CreateAlertRuleDtoConditions } from './createAlertRuleDtoConditions';

export interface CreateAlertRuleDto {
  /** Rule name */
  name: string;
  /** Rule description */
  description?: string;
  /** Rule type */
  ruleType: CreateAlertRuleDtoRuleType;
  /** Alert severity */
  severity: CreateAlertRuleDtoSeverity;
  /** Rule conditions (JSON) */
  conditions: CreateAlertRuleDtoConditions;
  /** Enable rule */
  enabled?: boolean;
}