// frontend/src/api/generated/models/createAlertRuleDtoRuleType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */

/**
 * Rule type
 */
export type CreateAlertRuleDtoRuleType = typeof CreateAlertRuleDtoRuleType[keyof typeof CreateAlertRuleDtoRuleType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateAlertRuleDtoRuleType = {
  risk_score: 'risk_score',
  transaction_value: 'transaction_value',
  tag_match: 'tag_match',
  velocity: 'velocity',
  cluster_risk: 'cluster_risk',
} as const;