// frontend/src/api/generated/models/createAlertRuleDtoSeverity.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */

/**
 * Alert severity
 */
export type CreateAlertRuleDtoSeverity = typeof CreateAlertRuleDtoSeverity[keyof typeof CreateAlertRuleDtoSeverity];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateAlertRuleDtoSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;